add leaf addition and removal and various int64 to int16 fixes

+++ phylo2vec/utils/random.py
@@ -1,4 +1,5 @@
 """Random utility functions: sampling and seeding."""
+
 import os
 import random
 import numpy as np
@@ -12,7 +13,9 @@ def sample(n_leaves):
     numpy.ndarray
         Phylo2Vec vector where v_i in {0, 1, ..., 2*i}
     """
-    return np.array([random.randint(0, 2 * i) for i in range(n_leaves - 1)])
+    return np.array(
+        [random.randint(0, 2 * i) for i in range(n_leaves - 1)], dtype=np.uint16
+    )
 
 
 def seed_everything(seed):
